# src/main/resources/api.yaml
openapi: 3.0.0
info:
  title: Math Functions API
  version: 1.0.0
  description: API для работы с математическими функциями
servers:
  - url: http://localhost:8080/api
    description: Local server
security:
  - basicAuth: []
paths:
...
  /users/register:
    post:
      summary: Регистрация нового пользователя
      tags: [Users]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                email:
                  type: string
      responses:
        '201':
          description: Пользователь создан
        '400':
          description: Неверные данные
        '409':
          description: Пользователь уже существует
  /users/me:
    get:
      summary: Получение информации о текущем пользователе
      tags: [Users]
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Не авторизован
  /users/{id}:
    get:
      summary: Получение пользователя по ID
      tags: [Users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: Пользователь не найден
    put:
      summary: Обновление пользователя
      tags: [Users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: Пользователь не найден
    delete:
      summary: Удаление пользователя
      tags: [Users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Удалено
        '404':
          description: Пользователь не найден
  /users/{id}/role:
    put:
      summary: Изменение роли пользователя
      tags: [Users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: role
          in: query
          required: true
          schema:
            type: string
            enum: [USER, ADMIN]
      responses:
        '200':
          description: Роль изменена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: Пользователь не найден
  /users:
    get:
      summary: Поиск пользователей
      tags: [Users]
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: role
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserShortResponse'

  # ОПЕРАЦИИ С ФУНКЦИЯМИ
...
components:
  schemas:
    UserResponse:
      type: object
      properties:
        userId:
          type: integer
        username:
          type: string
        role:
          type: string
        createdAt:
          type: string
          format: date-time
    UserShortResponse:
      type: object
      properties:
        userId:
          type: integer
        username:
          type: string
        role:
          type: string
    UserUpdateRequest:
      type: object
      properties:
        username:
          type: string
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: Basic authentication with username and password